% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{barplotResults}
\alias{barplotResults}
\title{Barplot of enriched results}
\usage{
barplotResults(
  object,
  x = "GeneRatio",
  color = "p.adjust",
  showCategory = 10,
  terms = NULL,
  font.size = 12
)
}
\arguments{
\item{object}{enrichResult or compareClusterResult object}

\item{x}{x axis, GeneRatio or Count}

\item{color}{color, one of pvalue, qvalue or p.adjust}

\item{showCategory}{number of terms to show}

\item{terms}{name of terms to show}

\item{font.size}{fontsize}
}
\description{
Barplot of enriched results
}
