% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{degPlotCluster}
\alias{degPlotCluster}
\title{plot the results of DE gene expression pattern}
\usage{
degPlotCluster(
  table,
  time,
  color = NULL,
  points = TRUE,
  boxes = TRUE,
  smooth = TRUE,
  lines = TRUE,
  facet = TRUE,
  facet_col = 4,
  facet_scales = "free_x",
  angle = 45
)
}
\arguments{
\item{table}{input data.frame}

\item{time}{the times or other conditions}

\item{color}{column treated as color}

\item{points}{whether plot points}

\item{boxes}{whether plot box}

\item{smooth}{whether smooth}

\item{lines}{whether plot the lines}

\item{facet}{where use facet}

\item{facet_col}{how many columns of figure}

\item{facet_scales}{facet scales}

\item{angle}{x axis lable angle}
}
\description{
plot the results of DE gene expression pattern
}
