% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{dotplotResults}
\alias{dotplotResults}
\title{Dotplot of enriched results}
\usage{
dotplotResults(
  object,
  x = "GeneRatio",
  color = "p.adjust",
  showCategory = 10,
  terms = NULL,
  size = "Count",
  font.size = 12
)
}
\arguments{
\item{object}{enrichResult or compareClusterResult object}

\item{x}{x axis, GeneRatio or Count}

\item{color}{color, one of pvalue, qvalue or p.adjust}

\item{showCategory}{number of terms to show}

\item{terms}{name of terms to show}

\item{size}{dot size}

\item{font.size}{fontsize}
}
\description{
Dotplot of enriched results
}
