% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{download.GEO}
\alias{download.GEO}
\title{Download GEO Datasets or supplementary files}
\usage{
download.GEO(geoID, out_dir = paste0(getwd(), "/GEO_Download"))
}
\arguments{
\item{geoID}{GEO Accession Number}

\item{out_dir}{where the files will download to}
}
\description{
Download GEO Datasets or supplementary files
}
