% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_data.R
\name{fortify_results}
\alias{fortify_results}
\title{convert enrichResult or compareClusterResult to a data.frame that ready for plot}
\usage{
fortify_results(model, showCategory = 10, terms = NULL, by = "geneRatio")
}
\arguments{
\item{model}{'enrichResult' or 'compareClusterResult' object}

\item{showCategory}{Category numbers to show}

\item{terms}{enriched terms to show}

\item{by}{one of Count and GeneRatio}
}
\value{
data.frame
}
\description{
convert enrichResult or compareClusterResult to a data.frame that ready for plot
}
