% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runing_functions.R
\name{get.DEGs}
\alias{get.DEGs}
\title{Get A List Of Differentially Expressed Genes}
\usage{
get.DEGs(dds, ctrl, treat, p.adjust = 0.05, abs.lfc = 1, save = FALSE)
}
\arguments{
\item{dds}{object of produced by DESeq}

\item{ctrl}{Control group}

\item{treat}{TreatMent groups}

\item{p.adjust}{Adjusted P-values}

\item{abs.lfc}{Log2 FoldChange Abs values}

\item{save}{save DEG List to Local Directory}
}
\value{
DEG List
}
\description{
Get A List Of Differentially Expressed Genes
}
