% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggtableResults}
\alias{ggtableResults}
\title{ggTable of enriched results}
\usage{
ggtableResults(
  object,
  by = "GeneRatio",
  color = "p.adjust",
  showCategory = 10,
  terms = NULL,
  font.size = 10
)
}
\arguments{
\item{object}{enrichResult or compareClusterResult object}

\item{by}{used as x axis and sorted by this value, c('GeneRatio', 'Count')}

\item{color}{color, one of pvalue, qvalue or p.adjust}

\item{showCategory}{number of terms to show}

\item{terms}{name of terms to show}

\item{font.size}{fontsize}
}
\description{
ggTable of enriched results
}
