% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runing_functions.R
\name{load.DEGs}
\alias{load.DEGs}
\title{Loadding DEGs from local directory}
\usage{
load.DEGs(filesName)
}
\arguments{
\item{filesName}{FileNames}
}
\value{
DEG List
}
\description{
Loadding DEGs from local directory
}
