% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runing_functions.R
\name{load_REGs}
\alias{load_REGs}
\title{Loadding REGs from local directory}
\usage{
load_REGs(filesName)
}
\arguments{
\item{filesName}{FileNames}
}
\value{
REG List
}
\description{
Loadding REGs from local directory
}
