% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainServer.R
\name{mainServer}
\alias{mainServer}
\title{create the mainly server functions}
\usage{
mainServer(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
create the mainly server functions
}
