% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotCorScatter}
\alias{plotCorScatter}
\title{scatter plot of two samples or two group samples with correlation}
\usage{
plotCorScatter(
  object,
  corr_group1,
  corr_group2,
  corr_method,
  design,
  digits = 4,
  pt.size,
  pt.alpha,
  pt.col
)
}
\arguments{
\item{object}{DESeqDataSet or DESeqTransform, if use DESeqDataSet log2(normalized.counts + 1) values will be used, if use DESeqTransform rlog or vst transformed values will be used}

\item{corr_group1}{group1, can be a single sample name or a condition name of a group samples}

\item{corr_group2}{group2, can be a single sample name or a condition name of a group samples}

\item{corr_method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}

\item{design}{design formula when runing DESeq}

\item{digits}{integer indicating the number of decimal places (round) or significant digits (signif) to be used. Negative values are allowed (see ‘Details’).}

\item{pt.size}{point size in plot}

\item{pt.alpha}{point alpha}

\item{pt.col}{point color}
}
\description{
scatter plot of two samples or two group samples with correlation
}
