% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{preview_GEO}
\alias{preview_GEO}
\title{Read in and preprocessing download files}
\usage{
preview_GEO(
  fetched_res,
  geoID,
  geo_files,
  genes_column,
  reads_column,
  header,
  row_names
)
}
\arguments{
\item{fetched_res}{object of geo.Download}

\item{geoID}{GEO Accession Number}

\item{geo_files}{a vector of input file names}

\item{genes_column}{the column number will used as gene names}

\item{reads_column}{the column number will used as reads counts of samples}

\item{header}{whethere use first row as header}

\item{row_names}{whethere use first column as rownames}
}
\value{
data.frame
}
\description{
Read in and preprocessing download files
}
