% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{publish_gostdot}
\alias{publish_gostdot}
\title{gprofiler2.dotplot of enriched results}
\usage{
publish_gostdot(
  object,
  by = "precision",
  color = "p_value",
  source = NULL,
  showCategory = 10,
  terms = NULL,
  font.size = 12
)
}
\arguments{
\item{object}{enrichResult or compareClusterResult object}

\item{by}{x axis, and order terms by this value}

\item{color}{color, one of pvalue}

\item{source}{DataSet source}

\item{showCategory}{number of terms to show}

\item{terms}{name of terms to show}

\item{font.size}{fontsize}
}
\description{
gprofiler2.dotplot of enriched results
}
