% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{publish_gostplot2}
\alias{publish_gostplot2}
\title{Plotting enriched results of gprofiler2, this function was modified from grofiler2's publish_gostplot function}
\usage{
publish_gostplot2(
  p,
  highlight_terms = NULL,
  filename = NULL,
  width = NA,
  fontsize = 14,
  height = NA,
  show_columns = c("source", "term_name", "term_size"),
  show_link = TRUE
)
}
\arguments{
\item{p}{ggplot object from gostplot}

\item{highlight_terms}{terms to highlight}

\item{filename}{filename to save}

\item{width}{plot width in inches}

\item{fontsize}{fontsize}

\item{height}{plot height in inches}

\item{show_columns}{additional coumns to show}

\item{show_link}{show the link of gProfiler websit in the bottom}
}
\description{
Plotting enriched results of gprofiler2, this function was modified from grofiler2's publish_gostplot function
}
