% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{publish_gosttable2}
\alias{publish_gosttable2}
\title{Plotting enriched results of gprofiler2, this function was modified from grofiler2's publish_gosttable function}
\usage{
publish_gosttable2(
  gostres,
  highlight_terms = NULL,
  use_colors = TRUE,
  show_link = TRUE,
  show_columns = c("source", "term_name", "term_size", "intersection_size"),
  filename = NULL,
  ggplot = TRUE,
  fontsize = 15
)
}
\arguments{
\item{gostres}{enriched results of gprofiler2}

\item{highlight_terms}{terms to highlight}

\item{use_colors}{whether use color to highlight p-values}

\item{show_link}{show the link of gProfiler websit in the bottom}

\item{show_columns}{names of additional columns to show}

\item{filename}{file name to create on disk and save the annotated plot. Filename extension should be from c("png", "pdf", "jpeg", "tiff", "bmp")}

\item{ggplot}{whether use ggplot}

\item{fontsize}{fontsize}
}
\description{
Plotting enriched results of gprofiler2, this function was modified from grofiler2's publish_gosttable function
}
