% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{read_files}
\alias{read_files}
\title{Function to import files by automatically identify delimiters}
\usage{
read_files(
  file_path = NULL,
  header = TRUE,
  row_names = 1,
  stringsAsFactors = F
)
}
\arguments{
\item{file_path}{the name of the file which the data are to be read from.}

\item{header}{a logical value indicating whether the file contains the names of the variables as its first line.}

\item{row_names}{a vector of row names.}

\item{stringsAsFactors}{logical: should character vectors be converted to factors?}
}
\value{
data.frame
}
\description{
Function to import files by automatically identify delimiters
}
