% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runing_functions.R
\name{remove.Batch}
\alias{remove.Batch}
\title{Remove batch effects from data}
\usage{
remove.Batch(
  expr.data,
  designTable,
  batch = "batch",
  batch2 = NULL,
  design = "condition",
  method = "ComBat"
)
}
\arguments{
\item{expr.data}{expression data matrix}

\item{designTable}{experiment design table}

\item{batch}{key words for batch factors}

\item{design}{unknown, conditions ?}

\item{method}{methods used to correct batch effects}
}
\description{
Remove batch effects from data
}
