% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runing_functions.R
\name{remove_Batch}
\alias{remove_Batch}
\title{Remove batch effects from data}
\usage{
remove_Batch(
  expr.data,
  designTable,
  key_words = "batch",
  design = "condition",
  method = "ComBat"
)
}
\arguments{
\item{expr.data}{expression data matrix}

\item{designTable}{experiment design table}

\item{key_words}{key words for batch factors}

\item{design}{unknown, conditions ?}

\item{method}{methods used to correct batch effects}
}
\description{
Remove batch effects from data
}
