% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ridgeplot2}
\alias{ridgeplot2}
\title{ridgeplot for gsea enriched results (modified from enrichplot::ridgeplot)}
\usage{
ridgeplot2(
  x = NULL,
  showCategory = 30,
  fill = "p.adjust",
  core_enrichment = TRUE,
  label_format = 30,
  orderBy = "NES",
  decreasing = FALSE,
  terms = NULL
)
}
\arguments{
\item{x}{gseaResults object}

\item{showCategory}{top n most significant terms}

\item{fill}{color to fill the density plot}

\item{core_enrichment}{whether only using core_enriched genes}

\item{label_format}{a numeric value sets wrap length, alternatively a custom function to format axis labels. by default wraps names longer that 30 characters}

\item{orderBy}{order the terms by}

\item{decreasing}{logical}

\item{terms}{term ID that of interest}
}
\description{
ridgeplot for gsea enriched results (modified from enrichplot::ridgeplot)
}
