% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runing_functions.R
\name{Run_DESeq2}
\alias{Run_DESeq2}
\title{Prepare and Run DESeq}
\usage{
Run_DESeq2(
  expr.data,
  sample.data,
  formula,
  test,
  fitType,
  sfType,
  betaPrior,
  reduced,
  minReplicatesForReplace,
  modelMatrixType,
  useT,
  minmu
)
}
\arguments{
\item{expr.data}{expression matrix}

\item{sample.data}{sample infomation meta data}

\item{formula}{experiment design formula}

\item{test}{either "Wald" or "LRT", which will then use either Wald significance tests (defined by nbinomWaldTest),
or the likelihood ratio test on the difference in deviance between a full and reduced model formula (defined by nbinomLRT)}

\item{fitType}{either "parametric", "local", or "mean" for the type of fitting of dispersions to the mean intensity.}

\item{sfType}{either "ratio", "poscounts", or "iterate" for teh type of size factor estimation.}

\item{betaPrior}{whether or not to put a zero-mean normal prior on the non-intercept coefficients}

\item{reduced}{for test="LRT", a reduced formula to compare against}

\item{minReplicatesForReplace}{the minimum number of replicates required in order to use replaceOutliers on a sample.}

\item{modelMatrixType}{either "standard" or "expanded", which describe how the model matrix, X of the GLM formula is formed.}

\item{useT}{logical, passed to nbinomWaldTest, default is FALSE, where Wald statistics are assumed to follow a standard Normal}

\item{minmu}{lower bound on the estimated count for fitting genewise dispersion and for use with nbinomWaldTest and nbinomLRT}
}
\description{
Prepare and Run DESeq
}
