% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{Search.GEO}
\alias{Search.GEO}
\title{Searching GEO Datasets by keywords}
\usage{
Search.GEO(
  db = "gds",
  organism = NULL,
  title = NULL,
  title_lg = NULL,
  description = NULL,
  description_lg = NULL,
  datasetType = "Expression profiling by high throughput sequencing",
  datasetType_lg = "AND",
  use_history = T
)
}
\arguments{
\item{db}{character, name of the database to search for.}

\item{organism}{character, organism to search for.}

\item{title}{character, keywords cantained in title}

\item{title_lg}{character, the logical connection between next search fields.}

\item{description}{character, keywords cantained in description}

\item{description_lg}{character, the logical connection between next search fields.}

\item{datasetType}{character, dataset type}

\item{datasetType_lg}{character, the logical connection between next search fields.}

\item{use_history}{logical. If TRUE return a web_history object for use in later calls to the NCBI}
}
\value{
data.frame
}
\description{
Searching GEO Datasets by keywords
}
