% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_anno_color}
\alias{set_anno_color}
\title{Auto set the annotation colors for pheatmap}
\usage{
set_anno_color(anno_row, anno_col)
}
\arguments{
\item{anno_row}{dataframe for annotate row colors}

\item{anno_col}{dataframe for annotate col colors}
}
\description{
Auto set the annotation colors for pheatmap
}
