% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{subset_Tab}
\alias{subset_Tab}
\title{Subset The Experiment design Table and keep the order}
\usage{
subset_Tab(dds, vars, selected)
}
\arguments{
\item{dds}{object of produced by DESeq}

\item{vars}{columns names}

\item{selected}{variable to keep}
}
\value{
Data.Frame
}
\description{
Subset The Experiment design Table and keep the order
}
