% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{upload.data}
\alias{upload.data}
\title{Load in data sets and creat dataframe of Reads Number}
\usage{
upload.data(path, header, row_names = TRUE)
}
\arguments{
\item{path}{the name of the file which the data are to be read from.}

\item{header}{a logical value indicating whether the file contains the names of the variables as its first line.}

\item{row_names}{a logical value indicating whether the file contains the names of the variables as its first column.}
}
\value{
data.frame
}
\description{
Load in data sets and creat dataframe of Reads Number
}
