% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load.REGs}
\alias{load.REGs}
\title{Loadding REGs from local directory}
\usage{
load.REGs(filesName)
}
\arguments{
\item{filesName}{FileNames}
}
\value{
REG List
}
\description{
Loadding REGs from local directory
}
